automatically backed up by learn

+++ lib/bachelor.rb
@@ -22,7 +22,7 @@ end
 def get_average_age_for_season(data, season)
   total_age = 0
   data[season].each do |contestant|
-    total_age += contestant['age'].to_i
+    total_age += contestant['age'].to_f
   end
   total_age/data[season].count
 end
